Change semaphore to use longs opposed to ints. Also allow permits to be different than 1

+++ precipice-core/src/main/java/net/uncontended/precipice/DefaultAsyncService.java
@@ -84,12 +84,12 @@ public class DefaultAsyncService implements AsyncService {
     }
 
     @Override
-    public int remainingCapacity() {
+    public long remainingCapacity() {
         return controller.remainingCapacity();
     }
 
     @Override
-    public int pendingCount() {
+    public long pendingCount() {
         return controller.pendingCount();
     }
